[refactor/#172] authAxios 이름 유지
