+++ src/lib/api/formCreate.ts
@@ -3,7 +3,7 @@
  */
 import { AxiosResponse } from 'axios'
 
-import authAxios from './axiosWithCredentials'
+import authAxios from './authAxios'
 
 export const postAlbaForm = async (
   requestBody: string,
